pdf-service/requirements.txt
weasyprint>=60.0
jinja2>=3.1.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
pyhanko>=0.19.0
flask>=3.0.0
